10/12 libft

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: sarbaill <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/03 17:19:40 by sarbaill          #+#    #+#             */
-/*   Updated: 2018/12/04 16:07:26 by sarbaill         ###   ########.fr       */
+/*   Updated: 2018/12/10 18:46:20 by sarbaill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int i;
-	char *str;
+	size_t	i;
+	char	*str;
 
 	i = 0;
 	str = (char*)s;
 	while (i < n)
 	{
 		if (str[i] == (char)c)
-			return(&str[i]);
+			return (&str[i]);
 		i++;
 	}
 	return (NULL);
