+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: sarbaill <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/02 17:22:16 by sarbaill          #+#    #+#             */
-/*   Updated: 2018/12/03 00:05:46 by sarbaill         ###   ########.fr       */
+/*   Updated: 2018/12/10 18:48:55 by sarbaill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int			 i;
+	size_t		i;
 	char		*d;
 	const char	*s;
